deleting files for presentation to illustrate directory creation and minor bug fix

+++ src/authoring/model/WeaponData.java
@@ -30,7 +30,7 @@ public class WeaponData implements IReadableData {
 		return projectileName;
 	}
 	
-	public void setRange(int range) throws Exception{
+	public void setRange(double range) throws Exception{
 		if (range < 0){
 			throw new Exception("Range cannot be a negative number.");
 		}
